product service added

+++ src/main/java/com/fiscalapi/models/ProductTax.java
@@ -32,7 +32,7 @@ public class ProductTax extends BaseDto {
     /**
      * Tasa del impuesto. El valor debe estar entre 0.00000 y 1.000000 (ej: 0.160000).
      */
-    private BigDecimal rate;
+    private double rate;
 
     /**
      * Código del impuesto según el catálogo del SAT (valores: "001", "002", "003").
@@ -94,7 +94,7 @@ public class ProductTax extends BaseDto {
      *
      * @return la tasa del impuesto.
      */
-    public BigDecimal getRate() {
+    public double getRate() {
         return rate;
     }
 
@@ -103,7 +103,7 @@ public class ProductTax extends BaseDto {
      *
      * @param rate la tasa a establecer.
      */
-    public void setRate(BigDecimal rate) {
+    public void setRate(double rate) {
         this.rate = rate;
     }
